sched: remove one Samsung's modification

+++ include/linux/sched.h
@@ -1295,7 +1295,7 @@ struct task_struct {
 	struct llist_node wake_entry;
 	int on_cpu;
 	struct task_struct *last_wakee;
-	unsigned long wakee_flips;
+	unsigned int wakee_flips;
 	unsigned long wakee_flip_decay_ts;
 
 	int wake_cpu;
